requirements.txt
anthropic>=0.40.0
pandas
pytest
python-dotenv>=1.0.0
streamlit
